feat: reuse header actions in admin layout
